refactors all code to C instead of C++, this was done to enable struct initialisation in the .data = {0,0,0} style and motivated by us not using any c++ feature
