+++ src/service/service_if.h
@@ -42,23 +42,19 @@
 #include <stdint.h>
 #include "ble.h"
 
-namespace service {
+/**@brief Function for initializing the Services generated by Bluetooth Developer Studio.
+ *
+ *
+ * @return      NRF_SUCCESS on successful initialization of services, otherwise an error code.
+ */
+uint32_t bluetooth_init(void);
 
-    /**@brief Function for initializing the Services generated by Bluetooth Developer Studio.
-     *
-     *
-     * @return      NRF_SUCCESS on successful initialization of services, otherwise an error code.
-     */
-    uint32_t bluetooth_init(void);
+/**@brief Function for handling the Application's BLE Stack events.
+ *
+ * @details Handles all events from the BLE stack of interest to all Bluetooth Developer Studio generated Services.
+ *
+ * @param[in]   p_ble_evt  Event received from the BLE stack.
+ */
+void bluetooth_on_ble_evt(ble_evt_t const * p_ble_evt);
 
-    /**@brief Function for handling the Application's BLE Stack events.
-     *
-     * @details Handles all events from the BLE stack of interest to all Bluetooth Developer Studio generated Services.
-     *
-     * @param[in]   p_ble_evt  Event received from the BLE stack.
-     */
-    void bluetooth_on_ble_evt(ble_evt_t const * p_ble_evt);
-
-
-    bool test_notify(uint8_t* data);
-}
+bool test_notify(uint8_t* data);
